+++ projects/plugins/boost/app/assets/src/js/modules/image-size-analysis/store/isa-groups.ts
@@ -1,7 +1,7 @@
 import { derived } from 'svelte/store';
 import { z } from 'zod';
 import { jetpack_boost_ds } from '../../../stores/data-sync-client';
-import { imageData } from '../datasync-image-size-analysis';
+import { imageData } from './isa-data';
 
 const Group = z.object( {
 	name: z.string(),
